fix importNode in multiple response loadDef
I was trying to get away with importing HTML nodes into XML, which everyone except IE is OK with.
As a workaround, I try the nice way, then if an error is thrown I switch to parsing XML from a string with Sarissa. This is fine for string values, but will break any event handlers registered on HTML elements.