store reference to containing question in JME scope
I couldn't think of a better way of letting an HTML object created in JME get hold of the containing question.
Question.scope has a property `question` which is a reference back to the question object. This lets JME functions access the question, without appealing to global variables or passing a question object down the whole evaluation chain.