some unreachable code that disallows fractions when "answer must be an integer"
I had written an issue (#433) which asked for fractions to be marked incorrect when "answer must be an integer" is selected.
I've added that, but decided that if "allow the student to enter a fraction" is ticked, then you really want to allow fractions.
If that isn't ticked, then the student gets a "can't submit" error anyway, so this error message can never be shown, unless something else changes later on.
Oh well!