name synonyms only apply to ops and function applications
Prompted by someone wanting to use `length` as a variable name, which is reasonable.

Previously, name synonyms were applied at the tokenisation step, which meant variable names were translated, as well as operator and function names. That doesn't make any sense - the synonyms are only there to cover alternative names for common functions.
Now, `jme.synonyms` has been split into two dictionaries: `jme.opSynonyms` and `jme.funcSynonyms`. Operator names are translated as the token is read, and function names are translated in `jme.shunt`, as soon as we're sure a name token is being used to name a function.