use i18next for localisation
Until now, we used R.js for localisation. That library was abandoned a long time ago by its creator, and wasn't very sophisticated - it lacked things like plural interpolation.
This commit switches to [i18next](http://i18next.com/), and changes the way the translated strings are included - now, every language we have is included in the compiled package, and the preferred one is used at init. This could open the way to switching languages while the exam is running.
The translated strings are stored in .json files in the locales directory. It's now possible to export languages directly from POEditor without having to put any boilerplate around the file.