fix x+(-y) rule, and bracketing of complex multiplicand
The rule simplifying x+(-y) to x-eval(-y) should only be applied when y is a number.

I also spotted a typo in texifyWouldBracketOpArg - I missed a ! from the front of the test to decide whether to bracket a complex number, meaning it did the opposite of what it should have.