don't pull unary minus out of a multiplicand when it's on a complex number
This is causing me a headache now, and might well cause me more headaches in the future.

A compromise I came up with a while ago was to make sure that complex numbers never have negative real part. To achieve this, if we end up with a complex number with negative real part, we rewrite it to -eval(-z), adding an explicit unary minus operation.
There's another set of rules which pull unary minuses out of brackets, with the aim of cancelling them out when multiplying two negatives.
When multiplying a complex number with negative real part, this looks awkward. For example, the following happens:
  (-7+i)*(x+1)  is rewritten to -(7-i)*(x+1)
The first form looks more natural.
However, consider
  (-7+i)*(-7+i)  vs  (7-i)*(7-i)
Is it reasonable to rewrite to the second form? I don't think so.
These rules are currently in 'basic', meaning they're always applied. I might spin them into a distinct ruleset, so you can decide whether you want this to happen or not.