run html attached callbacks as a separate job
There seemed to be a race condition with the `onHTMLAttached` callbacks.
This commit makes the callbacks run in a separate job to the `makeHTML`
call. Practically, this means that the thread is unblocked between the
two jobs, so something obscure to do with the DOM can happen.

This fixes a problem with the JSXGraph extension, where it found a way
to put the axis labels in a weird order.