stepsShown is only true if the student showed steps
There are two reasons the steps can be displayed: the student clicked
the "show steps" button and whatever penalty was applied, or the
question's answers have been revealed.

Previously, part.stepsShown was set to true in both cases, so it was
impossible to distinguish the two cases when deciding whether to apply
the steps penalty. The marking boiled down to a delicately balanced race
condition to make sure the correct score was reported.

I clearly thought I'd fixed this a while ago by introducing
part.stepsOpen, but hadn't cleared up the logic. Now, stepsShown is only
set to true as a result of the student asking to show steps while the
question is answerable (i.e. correct answers not revealed).
revealAnswers now only sets stepsOpen, not stepsShown.