After document loads, check that all scripts have executed
If a required script isn't included, the requireScript system just never executes it - it has to assume that the script will be loaded later.
All scripts should be present by the time the document is ready, so this adds a hook to check that every script has executed.
The missing script will have no unexecuted dependencies, so we report only the first such script.

fixes #184, which was filed back when each script was in a separate file.