groups of questions
Questions are now organised in groups, so each group can pick its
questions separately. You could use this to offer a variety of similar
questions, only showing one to the student.

There's an option to show the names of question groups in the navigation
menu and in the score breakdown on the results page.

see numbas/editor#315