add correctAnswerStyle to numberentry; number formatting functions take a style argument
The numberentry part now has a "correctAnswerStyle" setting, which specifies how to format the answer displayed to the student.

The functions dpformat and sigformat now take an optional third parameter `style`, which specifies how to format the number.
There's a new function `formatnumber(n,style)` which formats the given number with the given style, without specifying a precision. It's a wrapper round `Numbas.math.niceNumber(n,{style:style})`.