add context to MathJax processing errors
When Numbas catches a MathJax processing error, it know looks at the
parents of the offending element to get some context.

The values of the attribute `data-jme-context-description` on each of
the element's parents are concatenated, to build a description of where
the error occurred.

For example, it might say "MathJax processing error in question 1 part a
prompt".

This also adds util.getLocalisedAttribute(elem,name), which looks first
for the attribute with the given name, then `localise-<name>`,
localising that value if it exists.

Fixes #468. Thanks to Bernhard von Stengel for suggesting this.