attributes of the form `eval-<name>` sub variables
DOMcontentsubvars now looks for attributes of the form `eval-<name>` on
any element it touches. For each such attribute, the value is run
through `jme.subvars`, and the resulting string is stores in the
attribute <name>.

You could use this to set a class or style attribute on an element based
on a variable, for example.

Fixes #449