Dictionary JME data type
This commit adds a new JME data type, `dict`. It represents a map from
string keys to values of any type.

Dictionary literals use the same delimiters as lists, because curly
braces are already used. For example:

[
    "a": 1,
    "b": "this",
    "c": [1,2,3]
]

The following is equivalent:

dict(
    "a": 1,
    "b:" "this",
    "c": [1,2,3]
)

If you mix key/value pairs and plain elements, the compiler will throw
an error.

`[]` will create an empty list, *not* an empty dictionary. User `dict()`
to create an empty dictionary.

To make working with JSON easier, strings can now be marked as "safe",
meaning that variables won't be substituted in when they're evaluated.
This allows you to use curly braces with impunity.

There's a `keypair` datatype, which assigns a value to a string key.
Maybe I can reuse it later on for named arguments to functions.

New JME functions:

* dict(keys)
* dict+dict merges dicts
* keys(dict) returns a list of the dict's keys
* values(dict) returns a list of the dict's values
* items(dict) returns a list of [key,value] pairs
* dict[key] returns the value corresponding to the key
* get(dict,key,default) returns dict[key] if it exists, otherwise default
* key in dict returns true if dict has the given key
* json_decode(json) parses a JSON string
* json_encode(x) stringifies the given value
* safe(str) marks a string as safe - variables aren't substituted in when evaluated
* len(dict) gives the number of items in dict
* split(str,delimiter) splits a string at occurrences of delimiter

Other changes:

* let(dict,expr) assigns a variable for each key in dict, then evaluates expr
* Texify the `in` operator properly
* Numbas.jme.wrapValue wraps non-array objects as dicts, unless a type hint is given.
* Numbas.util.escapeHTML now escapes < > " and ', as well as &