scope inheritance happens on access
Until now, whenever you create a JME Scope which inherits from a parent
scope, every variable/function/ruleset defined in that scope was copied
over into the new scope at the time of creation.

This is not efficient! It becomes very obvious when nesting `map` or
`filter` calls - a new scope is created each time the function is
called.

The new method is to just keep a reference to the parent scope.
Resolving a variable (or function or ruleset) involves climbing the
hierarchy of scopes until a match is found.

This entails some big breaking changes:
* You can't just use scope.variables[name] to get a variable - you must
  use scope.getVariable(name)
* Deleting a variable is also different: use scope.deleteVariable(name)
  (this just sets a flag saying the variable is deleted, rather than
   climbing up and deleting the variable in all the parent scopes, where
   it may have originally been defined)
* scope.rulesets doesn't work any more - use scope.allRulesets() to
  collect them together.

As far as I can tell, I've fixed all interactions with scopes in the
core runtime. I can't think of any extensions that do clever things with
scopes, but I'm going to check them next.

It's possible that I've written some questions which mess about with
scopes; I'll have to do some clever searching in the editor for those,
but I'd be amazed if anyone else had done anything like that.