ranges are only enumerated explicitly when necessary
Until now, when a (discrete) TRange is created, every element of the
range is enumerated and appended to the value list on creation.
This takes a long time when the range is very big, and you might not
even use the enumeration - `random(range)`, for example, can pick at
random using arithmetic.

Now, TRange.value is just the list [start,end,step_size], and ranges are
converted to lists when necessary - it's up to functions operating on
ranges to do this conversion.

Convert a list to a range using math.rangeToList, and get the number of
elements of a discrete list with math.rangeSize

fixes #467