resolve rulesets in dependency order
custom rulesets can be defined in terms of other custom rulesets, in
which case they need to be evaluated in the right order.

This drags the ruleset-compilation process out of the question creation
and into Numbas.jme.variables.makeRulesets, following the same pattern
as makeVariables.

A ruleset's dependencies are evaluated first, and circular references or
references to undefined sets are caught.