DOMcontentsubvars looks at nodeType first
DOMContentsubvars assumed that its `element` argument has a
`hasAttribute` method. This is not true for comment nodes.

This commit rewrites DOMcontentsubvars to first switch based on
`element.nodeType` - nodes of type 1 (element) and 3 (text) can be
modified, while any other value (including `undefined`, in the case that
element isn't a DOM node) is silently ignored.