Remove the JS marking methods
The default marking methods for every part are now implemented using
JME marking scripts, instead of in JavaScript.
This gets rid of the mark_builtin and validate methods, and
studentAnswerAsJME now returns the value of the `interpreted_answer`
note from the marking algorithm.

Default marking scripts are defined in individual files in the
`marking_scripts` directory - when the exam is compiled, these are
gathered into `Numbas.marking_scripts`.