gapfill parts now use a marking script
Lots of changes to the finalise_state algorithm to get gapfills to work:
you need to be able to "lift" the current state, so that you can apply
the state operations for a gap and then scale them down afterwards.

This also necessitated functions submit_part and mark_part, which apply
marking to the part with the given path, and return its feedback, along
with credit and validness. mark_part just runs the marking algorithm but
doesn't apply its results to the part, so you can make decisions about where and
when to apply it.