add apply_marking_script JME function, and fix lifts
apply_marking_script(name,studentAnswer,settings,marks) applies the script with
the given name and options. Returns a dictionary mapping note names to
{feedback, value, valid}.
Could use it to reuse a built-in marking script as part of a bigger
marking scheme.

When there is a lift in the marking feedback (for example, when
concat_feedback is used), the `end` feedback item only skips to the end
of the lift, it doesn't end all marking for the part.