move stuff to do with simplification rules to a new file
The simplification rules stuff is really a dependency of the JME display
code, and has many uses outside display, so this spins the
simplification and pattern-matching stuff out into a jme-rules file, and
a namespace Numbas.jme.rules.

To maintain backwards compatibility, references to the objects that were
in Numbas.jme.display are copied over.