put brackets around fraction exponents
when a number is rendered as a fraction, it can sometimes mess with the
order of operations - brackets are needed to make sure that the order is
the same as when it's rendered as a decimal.

For the ops which put brackets around numbers -- *, -u, /, ^ -- if the
number is rendered as a fraction, it's bracketed.

fixes #485