big restructure of parts so they can be standalone
In order to get the marking tests in the editor to work properly, and
avoid repeating loads of code, I decided it'd be better if the Part
objects could be created outside an exam.

Previously, the code to load part settings from XML, and resume from
SCORM storage if necessary, were in the Part object's constructor. Now,
the constructor does the minimum necessary to build a "blank" part, and
new methods `loadFromXML` and `resume` do that job. Now you can set up the
part using a different method if required.

The storage engine is now a property of the part, rather than using
the global `Numbas.store`.

A part can now run without an attached display object.

Multiple choice parts have changed in a breaking way: previously, the
settings and XML were permuted to match the shuffles for
choices/answers. Because we can't count on XML and SCORM storage being
present at the same time, that has to change. So now the
MultipleResponsePart object keeps everything in the initial order, and
it's up to the display code to show the options in the shuffled order.
Apart from anything else, this makes the marking script much less of a
headache!

(I have a longer-term plan that questions, and then exams, should be
 more modular, so you could have multiple exams in a page, or mark a
 question headlessly)