'basic' ruleset always begins by being turned on
A change to the way Numbas.jme.collectRuleset works:
When the 'set' argument is a string, 'basic' is inserted at the start of
the list of set names, so it's turned on by default and must be
explicitly turned off.

e.g. \simplify[collectnumbers]{...} will have both 'basic' and
'collectNumbers' turned on, but \simplify[!basic,collectnumbers]{...}
will only have 'collectNumbers' turned on.

I think I avoided this because I didn't want to have an explicit
reference to any ruleset names in the collectRulesets code, but it's
much more convenient for 'basic' to be on by default.