m_all works better in non-commutative mode
and m_number is no longer classed as an "end term". I'm not sure why it
was originally, because the intention is that "end terms" should be able
to remain unmatched at the end.

The change for m_all makes sure it matches all the terms it can, before
moving on to the next term in the rule.