all the part types mark standalone
In tests/standalone.html, I create instances of each of the part types
and mark them against answers set in JavaScript, without any display
code or even a Question object.

The `submit_part` function in the script for gapfill needs a
`scope.question.getPart` method, which at the moment I shim.

If it was just `scope.getPart`, would that be any better?