starting to make questions work standalone
Because a question can either load variables or generate them afresh,
I've had to introduce a signals system to make everything happen at the
right time.

One benefit of this is that it'll be easier to make variable generation
happen asynchronously.

At the moment, I've structured things so it shuold be possible to make a
question standalone, outside an exam, but I haven't tried it yet.
Questions work inside a normal exam, which is good.