get rid of answerList
answerList was sort of redundant, and one of the final hang-overs from
the very first version of Numbas.

Now, the process of answering a question looks like this:

* p.storeAnswer(answer) - Save the object `answer` as `p.stagedAnswer`
* p.setStudentAnswer() - `p.stagedAnswer` is used to set whatever
    "submitted answer" properties the part has.
* p.mark() - Mark the part, using `p.rawStudentAnswerAsJME()`, which
    should read the properties set by `setStudentAnswer`

Since this changes the way `storeAnswer` works, old themes will break.
This is worth doing because the answerList format was always hard
to understand.