rename window.ko to Knockout
to avoid a name clash with geogebra, which pollutes the global namespace
with loads of short variable names.

This doesn't remove window.ko, so any questions referring to it
shouldn't break. Questions shouldn't technically rely on knockout being
available anyway - it's an aspect of the theme, not the core Numbas
runtime.

fixes #492