flatten nested multiplications
for example, '2^3 * 2^3 * 2^3' is compiled as ((2^3 * 2^3) * 2^3).

This flattens out the nested multiplications, so the string of
multiplications can be considered as a whole.

The '*' texify code was written assuming this would be the case; I
suppose I never got round to actually doing it!

For evaluation, I think it's still better to have multiplication as a
binary op, so this flattening only happens inside texify, and doesn't
change the input tree.