fix SCORM suspend for TExpression
jme.display.treeToJME now takes an option "wrapexpressions", which
renders TExpression tokens in the form `expression("...")`, so they can
be correctly re-parsed.

The reason it doesn't do this by default is for things like subvars into
the answer for a mathematical expression part, where you want the
subexpression to look like a normal part of the tree.