Tighten up texing special numbers
fixes #509, and consolidates treatment of special numbers.

There's now a list Numbas.jme.display.specialNumbers, of objects
defining the TeX and JME code for special numbers. Those are currently
e, pi, and infinity.

texifyWouldBracketOpArg now doesn't bracket special numbers when they're
raised to a power. (Without this check, it assumed they'd be rendered as
fractions and so need brackets)