custom part type settings are stored in their own dict
The JSON data that defines an instance of a custom part type used to
store the values for each of the settings at the top level.

This could clash with one of the built-in properties for a part, e.g.
'stepsPenalty'.

So now setting values are stored in a dictionary 'settings'.