signalbox errors don't halt the scheduler
The old scheduler is used to run an exam. Questions are now stand-alone,
and use Numbas.schedule.SignalBox objects. When one of these fail, it
shouldn't modify global state, preventing anything else from running.

This changes the SignalBox to just reject promises when they throw an
error. The Exam object now halts the scheduler if any of its questions
has an error.