don't mark parent parts as dirty when step modified
The behaviour around submitting steps is quite hard to get my head
round. Submitting a step doesn't cause the parent part to be submitted -
that's correct, because they can be interacted with independently,
except that the total earned through steps affects the score awarded
for the parent.

However, until now, when a step is marked as dirty, i.e. the student has
changed their answer, the parent part was also marked dirty. That meant
that the score earned from steps wouldn't be shown until the parent part
was submitted.

This changes the behaviour, so steps don't set their parent parts as
dirty. When you just submit steps, and don't enter an answer to the
parent part, the score earned from steps is shown for the part, and "you
were awarded X marks for your answers to the steps" is the only entry in
the feedback.