load input options for custom part types
Many of the input methods now have options which change their behaviour,
so it's easier for the part type author to make an input work the way
they want.

The marking script can access the `input_options` dictionary. I wanted
this so I didn't have to reproduce the expression I wrote for
correctAnswer.