less gnomic error message when the marking algo fails
This changes the error message shown to the student when the part
marking algorithm fails to a very generic one.
The student doesn't need to know how the marking works - they just need
to know that it's gone wrong. Since they can't do anything to fix it,
the message just says "please report this".