don't sub variables when evaluating Rule conditions
We had a scope with x := 2.
When matching the simplification rule ?;n*?;m, n isa "number" and m isa
"number" on the expression `4*x`, `m isa "number"` would evaluate to
true, because the value of x was subbed in.

This commit adds an argument `noSubstitution` to jme.Scope.evaluate. If
true, then variable values are not substituted in from the scope.