concat_feedback uses finalised_state, not markingFeedback
the messages in markingFeedback are of the form shown to students, not
Numbas.marking.feedback_item.

This adds some tests to ensure that the errors that alerted me to this
bug don't happen again. We had a JME part inside a gapfill, and the
feedback from the JME gap wasn't being correctly applied to the parent
part.

Thanks to Joshua Capel for reporting this.