add sortAnswers option to gapfill parts
towards #551

if sortAnswers is turned on, then the student's answers to the gaps are
sorted in ascending order before being marked.

Because the gaps aren't necessarily marked against what's in their
corresponding input widget, the input widgets don't show feedback state
or a feedback icon. Ideally, they'd show the icon corresponding to the
other gap marked as if it was in that position, but the plumbing is just
too hard for me to get my head round.

This adds a JME function sort_destinations, which maps list indices to
the positions they end up in when the list is sorted.