store exam.timeSpent and exam.stop in SCORM suspend data
The time spent and end time are saved in the suspend data. This only
happens when the student does something like submit an answer or change
question, but it's better than nothing. There isn't time to save data
when the window closes. At least the answer looks reasonable now.

fixes #534