a variety of fixes to matrix entry
the new marking algorithm didn't look at the allowFractions setting
(oops!)

loadFromJSON divides precisionPartialCredit by 100, because in the .exam
it's a number 0-100, and in the runtime it's 0-1.

The precision checking for matrix entry follows that of number entry:
the maximum precision used in the student's answer is calculated, and
each cell is compared with the corresponding cell in the correct answer
to that precision. Then, a penalty is applied if the precision the
student used is not the required one.

There's now a warning message if the student does not give all of their
cells to the same precision (only shown if fractions are not allowed).