add util.matchNotationStyles and JME `matchnumber`
For the quantities extension, I need to be able to match a number at the
start of a string, and obtain the unmatched remainder.
util.cleanNumber and util.parseNumber expect to parse the whole string.

Now, util.matchNotationStyles tries to match a number at the start of a
string, and has a `mustMatchAll` option which requires that the whole
string is matched (up to surrounding whitespace). cleanNumber uses this
flag.

There's a JME function `matchnumber` which calls matchNotationStyles and
returns a list of the form [matched original, parsed number]

This commit also fixes findvars on `try` so it finds variables used in
the first and third arguments (the expression to try, and the expression
to run when there's an error.