Functions to add new operators to the parser
I wanted to add an 'as' operator for the quantities extension. Until
now, the list of operator names has been hardcoded.
Now there are functions jme.addBinaryOperator, jme.addPrefixOperator and
jme.addPostfixOperator.

The dictionaries storing properties for operators are scattered all over
the place, and global. Ideally, stuff to do with parsing should be a
property of a Scope object. That's issue #362