signalbox promises reject properly
SignalBox.on returns a promise, but applies a .catch to it internally so
it can mark all other signals as failed. This .catch now returns a
Promise.reject() so that whatever originally called .on can also catch
the error. (If you don't return anything from .catch, the promise is
marked as 'resolved' by default)