properly keep track of base marking script
Previously, built-in Parts had a `markingScript` property on creation,
which was overwritten if a custom marking algorithm was used.
Custom part types don't have this until `getDefinition` is called, which
happens after the standard loadFromXML/JSON function.

This commit tidies all that up by adding a method `baseMarkingScript`,
which should always return the base marking script for a part type. For
custom part types, this calls `getDefinition`.