properly deal with nesting in StatefulScope
When the first expression in a try() clause throws an error, the
StatefulScope was setting `this.new_state` to true, which meant that any
feedback produced up to that point was lost.

Instead, we need to keep track of nesting depth, and just decrement it
when an evaluation finishes or throws an error. A new state is begun if
the nesting depth is 0.

This fixes #569