max accuracy on rationalApproximation
The numberentry part can ask for "Infinity" accuracy on the correct
answer fraction when the min and max value are the same.
This adds a clamp to rationalApproximation to use accuracy at most 30.
It also adds a get-out to the while loop to stop at 100 items. If we get
that far, we're not getting any more accurate.