Fix compareTrees on monomials with negative degrees
I was naively converting the degree of a monomial to a number by
assuming it was a TNum token, forgetting about unary minus.

Now, the degree is returned as a subexpression, which is run through
compareTrees to decide which should go first.

This also adds a couple of simplification rules to cancelTerms to catch
cases with a leading unary minus.