halfway through big refactor of parser
I think everything to do with parsing is now a property of the
jme.Parser object.
jme.compile uses a jme.standardParser object, which is a persistent
instance of jme.Parser, rather than creating a new one each time it's
called. This is because the addOperator functions are now methods of
jme.Parser, rather than functions which affect global variables.

This is all with the aim of improving the pattern-matching syntax. I've
added a few pattern-matching specific operators to a parser called
jme.rules.patternParser.

There are still undocumented methods around the place, and trying to
maintain backwards compatibility has made some function signatures
quite awkward. There are one or two breaking changes, which I need to
document.

I'm just committing at this point because it's home time!