add m_func and m_op, and match lists
m_func(name,args) matches a function application if the name, as as
string, and the arguments, as a list, match the given patterns,
Likewise for m_op with operators.

In order for this to work, I've hacked in the most basic matching logic
for lists: two lists match if they have the same length, and each
element in the candidate's list matches the corresponding pattern in the
pattern's list.