constraint checking seems to work
When an identified name only occurs once in the sequence, treat it like
a normal name assignment.

This means that when you have an identified name as part of a
subexpression, it will capture all terms that are assigned that name,
instead of only the first.

e.g. in  `m_number * (?`+);=x + m_number * (?`+);=x`, the constraint
should only be applied once we're matching the addition. In the product,
the identification only occurs once.

Need to think more about if this is sensible - could there be a better
way of specifiying the scope of identified names?