improved logic for deciding when to bracket numbers in treeToJME
I realised that the convoluted logic to decide when to bracket numbers
in JME was really asking the question 'is an arithmetic operation
implied?'
For example, the number '2+3i' is stored as a single number token, but
is rendered as a sum, so should be treated as one.
This allows us, once we've established the topmost implied operator, to
use the existing flags about when nested operations should be bracketed.