fixing the order of some exam init/load jobs
The exam's display was being initialised before percentScore
was calculated.

It turned out that jobs in the Exam.load sequence were using the
old synchronous job system instead of waiting for the 'question list
initialised' signal.

This fixes that, and adds a 'ready' signal to Exams, which the
start-exam stuff waits for, before initialising the display

fixes #578