add numerical_compare JME function
wraps the old jme.compare function.

The JME part marking script does the same thing, but in marking notes,
so it can give feedback about which bit failed.

I wanted this for pattern-matching so I can tell if a particular term is
equivalent to 1, for example, without doing it symbolically.