error while doing adaptive marking just causes 'incorrect'
I had a question which used adaptive marking to replace a parameter for
an algorithm. Certain values of parameters are inconsistent with the
algorithm's assumptions, so using them causes an error.
I should be catching those invalid values and giving better feedback,
but if I don't, Numbas shouldn't just die.
Showing the error message to the student isn't helpful - it's often
something inscrutable, like "list index error".

So now, if an error occurs while marking a part with "always replace
variables", the following happens:

* the part is marked incorrect.
* the student is told "This part could not be marked using your answers
  to previous parts", as a sort of hint that something further up is
  incorrect (they might've lost marks, but if the referenced part has
  zero credit that won't be the case)
* the error message is logged to the console, where the question author
  can find it