"show cell answer state" option for MCQ; only show highlight when not dirty
Question authors can choose not to show the correct/incorrect state on
individual choice inputs for MCQ parts. This is appropriate when the
marking is based on the entire choice matrix, as opposed to the default
where each choice adds or subtracts marks when ticked.

This also fixes #581, removing the feedback styling when a part is dirty