fix setting of studentanswerrows/columns in matrixentry part
When a matrix entry part is a gap, the submit_part call in the gapfill
part's marking algorithm loses the rows/columns information attached to
the student matrix.

It's not really a matrix, but a 2D array of strings, so I shouldn't have
been depending on that information being kept.

Instead, the MatrixEntryPart works out the size by looking at the
dimensions of the array.