getCorrectAnswer is called by mark_answer
the marking script JME function 'mark_part' calls Part.mark_answer
directly, instead of markAgainstScope, because it wants to get the 'raw'
marking results for the part, instead of a finalised set of feedback,
warnings and credit.

Previously, Part.getCorrectAnswer was only called in markAgainstScope.
It should be called before getting the marking parameters, and in fact
the marking algorithm should be evaluated against the scope the correct
answer is obtained from.

So, this changes Part.mark_answer so it gets passed a scope, and calls
getCorrectAnswer before evaluating the marking script.