Improve matching of - and /
Addition/subtraction and Multiplication/division are treated as the same
'class' of ops, in non-strict associative mode: Sums are matched as a
series of terms joined by '+', and products are matched as a series of
terms joined by '*'. Subtractions `a-b` are rewritten to `a+(-b)`, and
divisions `a/b` to `a*(/b)`, where `/` is a unary division operator.