lots of pattern-matching thinking
m_name, m_number and m_nothing are now $v, $n and $z.

I've been trying to fix the 'identified name' operator, ;=.
There's an issue of scope: do you want to catch repetitions of the same
thing inside one pattern term with a quantifier, e.g.
    (?;=x)`+ + $z
should match x*x*x but not x*y, but that identification shouldn't be
tested against any other terms in the wider pattern.

Or you do want to identify bits captured in different terms of the
pattern, e.g.
    $n * ?;=x + $n * ?;=x
should match 2x + 3x but not 2x + 3y.

The idea now is that there are 'inside' and 'outside' identified names.
Identified names are 'outside' until a quantifier is applied on top of
them, when they become 'inside'.

I should have kept track of test cases while doing this. At the moment,
the 'cancelTerms' rule doesn't work on "3*(-x) + 3*(-2)", because it
picks the 3 as the identified thing, and because -x isn't a number it
uses the default value of 1. I wonder if an m_exactly will fix this.