Returning a 'submit part' button for steps with one part
I've added a condition that returns the 'submit part' button for steps that have one part. This is necessary otherwise users are unable to easily submit steps that only have one part.

I have limited experience with xslt, so I don't suggest this is the best solution, but it does work.