JME part: remove variables defined in question from scope before evaluating expressions
When a variable used in the answer to a JME part is also defined in the
question, bad things could happen.

This adds an 'unset' JME function, which deletes the named variables,
functions and rulesets from the scope before evaluating its argument.
A dictionary giving lists of things defined by the question is also
passed into the part marking script.
Before evaluating a student or correct expression in a JME part, all
variables, functions and rulesets defined by the question are deleted.

There was also a related bug where the set of variables for which random
values are picked had the question's variables removed. Now, it's just
all free variables found in the expression.