question variables can be overridden by marking notes
We had a question with a variable 'rows', and a matrix entry part. The
marking algorithm for matrix entry parts defines a note 'rows',
representing the number of rows in the student's answer.

The question author should not need to know what notes are defined by
the marking algorithm. It's safe to assume that if a marking algorithm
defines a note with the same name as a question variable, it doesn't
care about the value of that variable.

So before a marking script is evaluated, the names of all notes are
marked as 'deleted' in the scope, so the marking script can't see any
variables that might already be defined with those names.