mark_part() uses Part.getScope()
Previously, the JME marking function `mark_part(part,answer)` passed the
scope it was being evaluated in to Part.mark_answer. When a marking
algorithm defines a variable with the same name as a question variable,
this causes problems!

The question author shouldn't need to know to avoid all names used in
marking algorithms, and marking algorithm authors shouldn't have to use
unusual variable names.

Now, it uses part.getScope() to get the scope the part is expecting to
use.

The issue I saw was the gapfill marking algorithm defining a bound
variable `gap` in a map over the gaps, and a question which defined a
numerical variable `gap`.